requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow>=10.0
whitenoise>=6.6
gunicorn>=21.0
python-decouple>=3.8
